First cut changes in preparation for environment upgrade (#2124)
* Avoid cubelists of cubelists.

* Replace np.product with np.prod.

* Replace np.int.

* Replace np.NAN with np.nan.

* Replace np.NAN with np.nan.

* Replace np.NaN with np.nan.

* Replace np.product with np.prod.

* Replace assertRaisesRegexp with assertRaisesRegex and use collections.abc.Callable instead of collections.Callable.

* Simplify test in test_flatten.py to avoid cubelist within a cubelist.

* Replace Cube(None) with an alternative.

* Unpin environments for testing.

* Remove pygam version.

* Update improver_a.yml

* Remove non-essential dependencies from yml files and add pins.

* Minor edits following review comments.

* Modify docstring to better reflect the inputs provided.

+++ improver_tests/threshold/conftest.py
@@ -80,7 +80,7 @@ def deterministic_cube() -> Cube:
 def single_realization_cube() -> Cube:
     """Return the diagnostic cube with a scalar realization coordinate."""
     cube = diagnostic_cube()
-    cube = add_coordinate(cube, [0], "realization", dtype=np.int)
+    cube = add_coordinate(cube, [0], "realization", dtype=np.int32)
     return cube
